pyproject.toml
[project]
name = "reference-api"
version = "2.0.0"
description = "Serves reference-repository JSON files as a REST API."
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "gitpython",
    "pydantic",
    "cachetools",
    "tomli; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint",
    "mypy",
    "types-cachetools",
    "httpx",
]

[tool.pylint.main]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-many-arguments",
    "too-many-positional-arguments",
]

[tool.setuptools.packages.find]
include = ["reference_api*"]
exclude = ["tests*"]
